chore(package): move files for requireable subcomponents
