+++ lib/add.js
@@ -1,7 +1,7 @@
 'use strict'
 
-var addOne = require('./utils/add-one')
-var addMany = require('./utils/add-many')
+var addOne = require('./helpers/add-one')
+var addMany = require('./helpers/add-many')
 
 module.exports = function add (objects) {
   var isArray = Array.isArray(objects)
