+++ lib/find-or-add.js
@@ -1,7 +1,7 @@
 'use strict'
 
-var findOrAddOne = require('./utils/find-or-add-one')
-var findOrAddMany = require('./utils/find-or-add-many')
+var findOrAddOne = require('./helpers/find-or-add-one')
+var findOrAddMany = require('./helpers/find-or-add-many')
 
 module.exports = function findOrAdd (idOrObjectOrArray, newObject) {
   var state = {
