+++ lib/find.js
@@ -1,7 +1,7 @@
 'use strict'
 
-var findOne = require('./utils/find-one')
-var findMany = require('./utils/find-many')
+var findOne = require('./helpers/find-one')
+var findMany = require('./helpers/find-many')
 
 module.exports = function find (objectsOrIds) {
   var state = {
