+++ lib/helpers/add-many.js
@@ -1,6 +1,6 @@
 'use strict'
 
-var toDoc = require('./to-doc')
+var toDoc = require('../utils/to-doc')
 
 module.exports = function addMany (state, objects) {
   var docs = objects.map(toDoc)
