+++ lib/helpers/add-one.js
@@ -1,6 +1,6 @@
 'use strict'
 
-var toDoc = require('./to-doc')
+var toDoc = require('../utils/to-doc')
 
 module.exports = function addOne (state, object) {
   if (typeof object !== 'object') {
