+++ lib/helpers/find-one.js
@@ -1,7 +1,7 @@
 'use strict'
 
-var toId = require('./to-id')
-var toObject = require('./to-object')
+var toId = require('../utils/to-id')
+var toObject = require('../utils/to-object')
 
 module.exports = function findOne (state, idOrObject) {
   var id = toId(idOrObject)
