+++ lib/helpers/find-or-add-many.js
@@ -1,4 +1,4 @@
-var toId = require('./to-id')
+var toId = require('../utils/to-id')
 var findMany = require('./find-many')
 var addMany = require('./add-many')
