+++ lib/helpers/find-or-add-one.js
@@ -1,4 +1,4 @@
-var toId = require('./to-id')
+var toId = require('../utils/to-id')
 var findOne = require('./find-one')
 var addOne = require('./add-one')
