+++ lib/helpers/update-many.js
@@ -2,9 +2,9 @@
 
 var extend = require('pouchdb-extend')
 
-var changeObject = require('./change-object')
-var toDoc = require('./to-doc')
-var toId = require('./to-id')
+var changeObject = require('../utils/change-object')
+var toDoc = require('../utils/to-doc')
+var toId = require('../utils/to-id')
 var findMany = require('./find-many')
 
 module.exports = function updateMany (state, array, change) {
