+++ lib/helpers/update-one.js
@@ -2,8 +2,8 @@
 
 var extend = require('pouchdb-extend')
 
-var changeObject = require('./change-object')
-var toDoc = require('./to-doc')
+var changeObject = require('../utils/change-object')
+var toDoc = require('../utils/to-doc')
 var findOne = require('./find-one')
 
 module.exports = function updateOne (state, idOrObject, change) {
