+++ lib/helpers/update-or-add-many.js
@@ -1,4 +1,4 @@
-var toId = require('./to-id')
+var toId = require('../utils/to-id')
 var addMany = require('./add-many')
 var updateMany = require('./update-many')
