+++ lib/helpers/update-or-add-one.js
@@ -1,4 +1,4 @@
-var toId = require('./to-id')
+var toId = require('../utils/to-id')
 var addOne = require('./add-one')
 var updateOne = require('./update-one')
