+++ lib/remove.js
@@ -1,7 +1,7 @@
 'use strict'
 
-var updateOne = require('./utils/update-one')
-var updateMany = require('./utils/update-many')
+var updateOne = require('./helpers/update-one')
+var updateMany = require('./helpers/update-many')
 
 module.exports = function remove (objectsOrIds) {
   var state = {
