+++ lib/update-or-add.js
@@ -1,7 +1,7 @@
 'use strict'
 
-var updateOrAddOne = require('./utils/update-or-add-one')
-var updateOrAddMany = require('./utils/update-or-add-many')
+var updateOrAddOne = require('./helpers/update-or-add-one')
+var updateOrAddMany = require('./helpers/update-or-add-many')
 
 module.exports = function updateOrAdd (idOrObjectOrArray, newObject) {
   var state = {
