+++ lib/update.js
@@ -1,7 +1,7 @@
 'use strict'
 
-var updateOne = require('./utils/update-one')
-var updateMany = require('./utils/update-many')
+var updateOne = require('./helpers/update-one')
+var updateMany = require('./helpers/update-many')
 
 module.exports = function update (objectsOrIds, change) {
   var Promise = this.PouchDB.utils.Promise
